perf: limit inbox prompt backfill concurrency
